reply hook
